+++ src/main/java/com/clova/anifriends/domain/shelter/controller/ShelterController.java
@@ -1,14 +1,14 @@
 package com.clova.anifriends.domain.shelter.controller;
 
 import com.clova.anifriends.domain.auth.LoginUser;
-import com.clova.anifriends.domain.shelter.dto.CheckDuplicateShelterEmailRequest;
-import com.clova.anifriends.domain.shelter.dto.CheckDuplicateShelterResponse;
-import com.clova.anifriends.domain.shelter.dto.FindShelterDetailResponse;
-import com.clova.anifriends.domain.shelter.dto.FindShelterMyPageResponse;
-import com.clova.anifriends.domain.shelter.dto.FindShelterSimpleByVolunteerResponse;
-import com.clova.anifriends.domain.shelter.dto.RegisterShelterRequest;
-import com.clova.anifriends.domain.shelter.dto.UpdateAddressStatusRequest;
-import com.clova.anifriends.domain.shelter.dto.UpdateShelterPasswordRequest;
+import com.clova.anifriends.domain.shelter.dto.request.CheckDuplicateShelterEmailRequest;
+import com.clova.anifriends.domain.shelter.dto.request.RegisterShelterRequest;
+import com.clova.anifriends.domain.shelter.dto.request.UpdateAddressStatusRequest;
+import com.clova.anifriends.domain.shelter.dto.request.UpdateShelterPasswordRequest;
+import com.clova.anifriends.domain.shelter.dto.response.CheckDuplicateShelterResponse;
+import com.clova.anifriends.domain.shelter.dto.response.FindShelterDetailResponse;
+import com.clova.anifriends.domain.shelter.dto.response.FindShelterMyPageResponse;
+import com.clova.anifriends.domain.shelter.dto.response.FindShelterSimpleByVolunteerResponse;
 import com.clova.anifriends.domain.shelter.service.ShelterService;
 import jakarta.validation.Valid;
 import java.net.URI;
