+++ src/main/java/com/clova/anifriends/domain/shelter/dto/response/FindShelterMyPageResponse.java
@@ -1,4 +1,4 @@
-package com.clova.anifriends.domain.shelter.dto;
+package com.clova.anifriends.domain.shelter.dto.response;
 
 import com.clova.anifriends.domain.shelter.Shelter;
