+++ src/test/java/com/clova/anifriends/domain/shelter/controller/ShelterControllerTest.java
@@ -22,14 +22,14 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import com.clova.anifriends.base.BaseControllerTest;
 import com.clova.anifriends.docs.format.DocumentationFormatGenerator;
 import com.clova.anifriends.domain.shelter.Shelter;
-import com.clova.anifriends.domain.shelter.dto.CheckDuplicateShelterEmailRequest;
-import com.clova.anifriends.domain.shelter.dto.CheckDuplicateShelterResponse;
-import com.clova.anifriends.domain.shelter.dto.FindShelterDetailResponse;
-import com.clova.anifriends.domain.shelter.dto.FindShelterMyPageResponse;
-import com.clova.anifriends.domain.shelter.dto.FindShelterSimpleByVolunteerResponse;
-import com.clova.anifriends.domain.shelter.dto.RegisterShelterRequest;
-import com.clova.anifriends.domain.shelter.dto.UpdateAddressStatusRequest;
-import com.clova.anifriends.domain.shelter.dto.UpdateShelterPasswordRequest;
+import com.clova.anifriends.domain.shelter.dto.request.CheckDuplicateShelterEmailRequest;
+import com.clova.anifriends.domain.shelter.dto.request.RegisterShelterRequest;
+import com.clova.anifriends.domain.shelter.dto.request.UpdateAddressStatusRequest;
+import com.clova.anifriends.domain.shelter.dto.request.UpdateShelterPasswordRequest;
+import com.clova.anifriends.domain.shelter.dto.response.CheckDuplicateShelterResponse;
+import com.clova.anifriends.domain.shelter.dto.response.FindShelterDetailResponse;
+import com.clova.anifriends.domain.shelter.dto.response.FindShelterMyPageResponse;
+import com.clova.anifriends.domain.shelter.dto.response.FindShelterSimpleByVolunteerResponse;
 import com.clova.anifriends.domain.shelter.support.ShelterFixture;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
